chore(): move code under src folder
