+++ src/reporters/console-reporter.js
