+++ src/reporters/string-reporter.js
