+++ src/space.test.js
@@ -1,4 +1,4 @@
-const { Metrics, InMemoryReporter } = require('./index');
+const { Metrics, InMemoryReporter } = require('../index');
 
 describe('Space', () => {
   describe('meter', () => {
